refactor: Refactor Axios Module Name & Dir
