+++ client/src/js/main.js
@@ -1,12 +1,12 @@
 import task from './task';
 import time from './time';
-import headerModalTurnOnOff from './turn-on-off-modal';
-import putSettings from './axios-header-modal';
+import modalFunc from './header-modal';
+import watchUpdate from './axios/update-settings';
 
 time();
 
-headerModalTurnOnOff();
+modalFunc();
 
 task();
 
-putSettings();
+watchUpdate();
